Move Input/Output to new commandio module
The types module should provide the core type definitions and this has
nothing to do with the types Input and Output which are used in the
interface of commands.
