+++ src/clientmanager.h
@@ -4,6 +4,7 @@
 #include <X11/X.h>
 #include <unordered_map>
 
+#include "commandio.h"
 #include "link.h"
 #include "object.h"
 #include "signal.h"
