+++ src/finite.h
@@ -3,6 +3,7 @@
 #include <type_traits>
 
 #include "completion.h"
+#include "types.h"
 
 /** The Finite<T> class defines the Converter methods
  * for enum class types, that is, for types with only finitely many values.
