+++ src/rectangle.h
@@ -1,7 +1,7 @@
 #ifndef __HERBSTLUFT_RECTANGLE_H_
 #define __HERBSTLUFT_RECTANGLE_H_
 
-#include "types.h"
+#include "commandio.h"
 #include "x11-types.h"
 
 RectangleVec disjoin_rects(const RectangleVec &buf);
