+++ src/tagmanager.h
@@ -1,6 +1,7 @@
 #ifndef __HLWM_TAGMANAGER_H_
 #define __HLWM_TAGMANAGER_H_
 
+#include "commandio.h"
 #include "byname.h"
 #include "indexingobject.h"
 #include "link.h"
