+++ src/types.cpp
@@ -4,35 +4,6 @@
 
 using std::string;
 
-Input &Input::operator>>(string &val)
-{
-    ArgList::operator>>(val);
-    return *this;
-}
-
-Input Input::fromHere()
-{
-    string cmd;
-    if (empty()) {
-        return {{}, {}};
-    }
-
-    return Input(*(begin()), Container(begin() + 1, end()));
-}
-
-void Input::replace(const string &from, const string &to)
-{
-    for (auto &v : *container_) {
-        if (v == from) {
-            v = to;
-        }
-
-        if (*command_ == from) {
-            *command_ = to;
-        }
-    }
-}
-
 template<> void Converter<bool>::complete(Completion& complete, bool const* relativeTo)
 {
     complete.full({ "on", "off", "true", "false" });
