+++ src/utils.h
@@ -9,6 +9,7 @@
 #include <sstream>
 #include <string>
 
+#include "commandio.h"
 #include "types.h"
 
 #define LENGTH(X) (sizeof(X)/sizeof(*(X)))
